+++ .idea/encodings.xml
@@ -7,7 +7,6 @@
     <file url="file://$PROJECT_DIR$/auth-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/auth/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/book-service/src/main/java" charset="UTF-8" />
-    <file url="file://$PROJECT_DIR$/book-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/book-service/src/main/resources" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/bookmark-service/src/main/java" charset="UTF-8" />
     <file url="file://$PROJECT_DIR$/chapter-service/src/main/java" charset="UTF-8" />
