+++ chapter-service/src/main/java/org/chapterservice/ChapterServiceApplication.java
@@ -1,0 +1,11 @@
+package org.chapterservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ChapterServiceApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(ChapterServiceApplication.class, args);
+    }
+}
