+++ chapter-service/src/main/java/org/chapterservice/mapper/ChapterMapper.java
@@ -1,0 +1,12 @@
+package org.chapterservice.mapper;
+
+import org.chapterservice.dto.request.ChapterRequest;
+import org.chapterservice.dto.response.ChapterResponse;
+import org.chapterservice.entity.Chapter;
+import org.mapstruct.Mapper;
+
+@Mapper(componentModel = "spring")
+public interface ChapterMapper {
+    Chapter toEntity(ChapterRequest request);
+    ChapterResponse toResponse(Chapter chapter);
+}
